refactor(019/extract-discover): move filesystem walker to discover.rs
Extracts the candidate-binary discovery code from binary/mod.rs
into a new sibling binary/discover.rs:

  pub(super) fn discover_binaries     (was private fn)
  fn walk_dir                         (private, unchanged)
  fn is_supported_binary              (private, unchanged)
  pub(crate) fn detect_format         (visibility unchanged)

Net: binary/mod.rs 1588 → 1510 LOC (-78). discover.rs is 89 LOC.
No tests live in this section — the existing tests/scan_binary.rs
integration test (1337 LOC) covers discovery end-to-end through
the public API.

Verified:
- cargo +stable check --workspace --tests passes clean
- All 27 byte-identity goldens unchanged: regen produces zero diff
- Removed unused `PathBuf` import in mod.rs (discover_binaries was
  the only consumer of the explicit import)

Per quickstart.md, this is the second of 4 atomic extraction
commits. Next: entry.rs (must come before scan.rs because
scan.rs imports BinaryScan from entry.rs).
